+++ hyperspy/tests/drawing/test_mpl_testing_setup.py
@@ -24,7 +24,7 @@ from hyperspy.misc.test_utils import check_running_tests_in_CI
 
 def test_mlp_agg_for_CI_testing():
     if check_running_tests_in_CI():
-        assert matplotlib.get_backend() == "agg"
+        assert matplotlib.get_backend().lower() == "agg"
 
 
 @pytest.fixture