+++ hyperspy/tests/drawing/test_plot_signal2d.py
@@ -34,6 +34,7 @@ except ImportError:
 import traits.api as t
 
 import hyperspy.api as hs
+from hyperspy.decorators import lazifyTestClass
 from hyperspy.drawing.utils import make_cmap, plot_RGB_map
 from hyperspy.tests.drawing.test_plot_signal import _TestPlot
 
@@ -607,6 +608,18 @@ def test_plot_images_tranpose():
     hs.plot.plot_images([a, b])
 
 
+def test_plot_images_update():
+    s = hs.signals.Signal2D(np.arange(100).reshape(10, 10))
+    s2 = s / 2
+    axs = hs.plot.plot_images([s, s2])
+
+    s.data = -s.data - 10
+    s.events.data_changed.trigger(s)
+
+    np.testing.assert_allclose(axs[0].images[0].get_array()[0, :2], [-10, -11])
+    np.testing.assert_allclose(axs[1].images[0].get_array()[0, :2], [0, 0.5])
+
+
 # Ignore numpy warning about clipping np.nan values
 @pytest.mark.filterwarnings("ignore:Passing `np.nan` to mean no clipping in np.clip")
 def test_plot_with_non_finite_value():
@@ -834,6 +847,129 @@ def test_plot_images_bool():
     s = hs.signals.Signal2D(data)
 
     hs.plot.plot_images(s)
+
+
+def test_plot_static_signal_nav():
+    s = hs.signals.Signal2D(np.ones((20, 20, 10, 10)))
+    nav = hs.signals.Signal2D(np.ones((20, 20)))
+    s.plot(navigator=nav)
+
+
+@lazifyTestClass
+class TestDynamicNavigatorPlot:
+    def setup_method(self, method):
+        self.signal5d2d = hs.signals.Signal2D(
+            np.arange((10**5)).reshape(
+                (
+                    10,
+                    10,
+                    10,
+                    10,
+                    10,
+                )
+            )
+        )
+        self.signal6d2d = hs.signals.Signal2D(
+            np.arange((10**6)).reshape((10, 10, 10, 10, 10, 10))
+        )
+        self.signal4d1d = hs.signals.Signal1D(
+            np.arange((10**4)).reshape((10, 10, 10, 10))
+        )
+        self.signal5d1d = hs.signals.Signal1D(
+            np.arange((10**5)).reshape(
+                (
+                    10,
+                    10,
+                    10,
+                    10,
+                    10,
+                )
+            )
+        )
+
+    def test_plot_5d(self):
+        import numpy as np
+
+        import hyperspy.api as hs
+
+        s = self.signal5d2d
+        nav = hs.signals.BaseSignal(np.arange((10 * 10 * 10)).reshape(10, 10, 10))
+        s.plot(navigator=nav)
+        data1 = s._plot.navigator_plot._current_data
+        s.axes_manager.indices = (0, 0, 1)
+        data2 = s._plot.navigator_plot._current_data
+        assert not np.array_equal(data1, data2)
+        s.axes_manager.indices = (0, 2, 1)
+        data3 = s._plot.navigator_plot._current_data
+        assert np.array_equal(data2, data3)
+
+    def test_plot_5d_2(self):
+        import numpy as np
+
+        import hyperspy.api as hs
+
+        s = self.signal5d2d
+        nav = hs.signals.Signal2D(np.arange((10 * 10 * 10)).reshape(10, 10, 10))
+        s.plot(navigator=nav)
+        data1 = s._plot.navigator_plot._current_data
+        s.axes_manager.indices = (0, 0, 1)
+        data2 = s._plot.navigator_plot._current_data
+        assert not np.array_equal(data1, data2)
+        s.axes_manager.indices = (0, 2, 1)
+        data3 = s._plot.navigator_plot._current_data
+        assert np.array_equal(data2, data3)
+
+    def test_plot_6d(self):
+        import numpy as np
+
+        import hyperspy.api as hs
+
+        s = self.signal6d2d
+        nav = hs.signals.BaseSignal(
+            np.arange((10 * 10 * 10 * 10)).reshape(10, 10, 10, 10)
+        )
+        s.plot(navigator=nav)
+        data1 = s._plot.navigator_plot._current_data
+        s.axes_manager.indices = (0, 0, 0, 1)
+        data2 = s._plot.navigator_plot._current_data
+        assert not np.array_equal(data1, data2)
+        s.axes_manager.indices = (0, 2, 0, 1)
+        data3 = s._plot.navigator_plot._current_data
+        assert np.array_equal(data2, data3)
+
+    def test_plot_4d_1dSignal(self):
+        import numpy as np
+
+        import hyperspy.api as hs
+
+        s = self.signal4d1d
+        nav = hs.signals.BaseSignal(np.arange((10 * 10 * 10)).reshape(10, 10, 10))
+        s.plot(navigator=nav)
+        data1 = s._plot.navigator_plot._current_data
+        s.axes_manager.indices = (0, 0, 1)
+        data2 = s._plot.navigator_plot._current_data
+        assert not np.array_equal(data1, data2)
+        s.axes_manager.indices = (0, 2, 1)
+        data3 = s._plot.navigator_plot._current_data
+        assert np.array_equal(data2, data3)
+
+    def test_plot_5d_1dsignal(self):
+        import numpy as np
+
+        import hyperspy.api as hs
+
+        s = self.signal5d1d
+        nav = hs.signals.BaseSignal(
+            np.arange((10 * 10 * 10 * 10)).reshape(10, 10, 10, 10)
+        )
+        s.plot(navigator=nav)
+        data1 = s._plot.navigator_plot._current_data
+        s.axes_manager.indices = (0, 0, 0, 1)
+        data2 = s._plot.navigator_plot._current_data
+        assert not np.array_equal(data1, data2)
+        s.axes_manager.indices = (0, 2, 0, 1)
+        data3 = s._plot.navigator_plot._current_data
+        assert np.array_equal(data2, data3)
 
 
 @pytest.mark.parametrize("axes_decor", ["all", "ticks", "off", None])